Fixes NTOS Bluescreen when in admin mode
Fixes a bluescreen caused by a nullkey when in admin mode.

NTOS bluescreens in the chat client because it attempts to render a mute button on a client after the admin mode has "removed" them from the room.

Added a null check to prevent this from happening, so it will check the type exists rather than just exploding.

Thanks to TG Station old prs for helping find this issue.